requirements.txt
streamlit==1.36.0
pandas>=2.2.3
numpy>=2.1.1
openpyxl>=3.1.5
boto3>=1.34.162
python-dotenv>=1.0.1
openai==0.28.1
python-docx==1.1.2
matplotlib>=3.9.0
